refactored default to/from array impl
